Migrate all of the SQLite building code to a different repo. (#3)
This allows the SQLite-building code to be used in other (non-BioC) contexts.
Now this repository is solely dedicated to building metadata indices for
Bioconductor, along with hosting of the schemas for BioC metadata.

We also take this opportunity to merge the scRNAseq schema into the BioC schema
under the 'takane' application, given that none of the metadata fields were
specific to the scRNAseq package anyway. This simplifies the setup for scRNAseq
(and future packages) by consolidating everything into one SQLite file.